quick sort

+++ sort.h
@@ -28,8 +28,8 @@ void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
 void swap(int *a, int *b);
-int partition(int *array, size_t low, size_t high, size_t size);
-void _qsort(int *array, size_t init, size_t final, size_t size);
+int partition(int *array, int low, int high, int size);
+void _qsort(int *array, int init, int final, int size);
 
 
 #endif
